Im/more tensorboard sleep (#94)
* delete train_wake_sleep.py as pretty deprecated by now

* added option to not overwrite folder (tensorboard takes care of managing multiple runs) in train.py

* small bug

* bug in decoder that was still passing in psf instead of psf_params

* bug related to number of bands in psf_params

* can now do multiple runs in same folder with train.py and automatically generate multiple version in tensorboard

* added correct checkpoint versoining

* added hparams logging

* logging hparams, but not yet tensorboard

* added plots for looking at reconstructed image and residuals of sleep phase

* added galaxy images to tensorboard

* unused imports and a comment

* made render_multiple_* private and made explicit shape assertion in render_image, sleep reconstructed image should work

* galaxy_bool in map_estimate is returned with correct dimensions.

* * actually kept flexibility (with both renders) but made it explicit by adding max_sources to render_image
* if statement to avoid error when plotting empty scenes in sleep.py reconstruction
* go back to render_multiple_stars in wake.py since want to add background.
* test_galaxy_encoder.py uses 5 images to avoid error when plotting reconstructions

* should pass on cpu now , added option to not do validation plots until runtime is improved

* assert to avoid confusion with axis indexing

* consistent private functions in decoder

* added validation_plot_start for burn-in period of plotting validation plots

* still plot something when no sources were detected.

* actually doing it now

+++ bin/utils.py
@@ -28,7 +28,7 @@ def add_path_args(parser):
     return parser
 
 
-def setup_paths(args):
+def setup_paths(args, enforce_overwrite=True):
     root_path = Path(args.root)
     paths = {
         "root": root_path,
@@ -37,10 +37,11 @@ def setup_paths(args):
         "output": root_path.joinpath(f"logs/{args.output}"),
     }
 
-    assert not paths["output"].exists() or args.overwrite, "Enforcing overwrite."
-    if args.overwrite and paths["output"].exists():
-        shutil.rmtree(paths["output"])
-    paths["output"].mkdir(parents=False)
+    if enforce_overwrite:
+        assert not paths["output"].exists() or args.overwrite, "Enforcing overwrite."
+        if args.overwrite and paths["output"].exists():
+            shutil.rmtree(paths["output"])
+    paths["output"].mkdir(parents=False, exist_ok=not enforce_overwrite)
 
     for p in paths.values():
         assert p.exists(), f"path {p.as_posix()} does not exist"
